Conflicts:
	Makefile.in
	addons/skin.confluence/720p/IncludesHomeMenuItems.xml
	addons/skin.confluence/language/English/strings.xml
	xbmc/Application.cpp

+++ lib/ffmpeg/build_xbmc_win32.sh
@@ -1,14 +1,26 @@
 #!/bin/bash
 
-if [ -d .libs ]
+MAKEFLAGS=""
+
+if [ "$1" == "clean" ]
 then
-rm -r .libs
+  if [ -d .libs ]
+  then
+    rm -r .libs
+  fi
+  make distclean
 fi
 
-if [ -f config.mak ]
-then
-make distclean
+if [ $NUMBER_OF_PROCESSORS > 1 ]; then
+  MAKEFLAGS=-j$NUMBER_OF_PROCESSORS
 fi
+
+if [ ! -d .libs ]; then
+  mkdir .libs
+fi
+
+# add --enable-debug (remove --disable-debug ofc) to get ffmpeg log messages in xbmc.log
+# the resulting debug dll's are twice to fourth time the size of the release binaries
 
 OPTIONS="
 --enable-shared \
@@ -32,18 +44,17 @@ OPTIONS="
 --enable-encoder=ac3 \
 --enable-encoder=aac \
 --enable-runtime-cpudetect \
---disable-debug \
+--enable-avfilter \
 --disable-doc"
 
 ./configure --extra-cflags="-fno-common -Iinclude-xbmc-win32/dxva2" --extra-ldflags="-L/xbmc/system/players/dvdplayer" ${OPTIONS} &&
  
-make && 
-mkdir .libs &&
+make $MAKEFLAGS &&
 cp lib*/*.dll .libs/ &&
-mv .libs/swscale-0.dll .libs/swscale-0.6.1.dll &&
 cp .libs/avcodec-52.dll /xbmc/system/players/dvdplayer/ &&
 cp .libs/avcore-0.dll /xbmc/system/players/dvdplayer/ &&
 cp .libs/avformat-52.dll /xbmc/system/players/dvdplayer/ &&
 cp .libs/avutil-50.dll /xbmc/system/players/dvdplayer/ &&
+cp .libs/avfilter-1.dll /xbmc/system/players/dvdplayer/ &&
 cp .libs/postproc-51.dll /xbmc/system/players/dvdplayer/ &&
-cp .libs/swscale-0.6.1.dll /xbmc/system/players/dvdplayer/
+cp .libs/swscale-0.dll /xbmc/system/players/dvdplayer/
